Start the modules lesson and auto compile ts
Managed to make an auto compile for TypeScript files using an extension
called File Watcher, that allows the user to set a command to run
whenever a file is edited. This will help me to make and test scripts
faster. Also, I started using the Live Server extension to test the
changes in the pages without needing to refresh the browser manually.